LayoutTests/crypto/webkitSubtle/aes-cbc-unwrap-failure-expected.txt

CONSOLE MESSAGE: line 23: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test unwrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing an unwrapping key...
PASS unwrappingKey.algorithm.name is 'AES-CBC'
Unwrapping a key...
PASS Promise rejected
PASS successfullyParsed is true

TEST COMPLETE
